moved restore to a file so it can be used in fire event
