+++ lib/delegate.js
@@ -1,6 +1,7 @@
 'use strict'
 const attach = require('./attach')
 const vstamp = require('vigour-stamp')
+const restore = require('./restore')
 
 module.exports = function delegateEvent (key, e) {
   var target = e.target
@@ -13,20 +14,8 @@ module.exports = function delegateEvent (key, e) {
         if (!stamp) {
           stamp = vstamp.create(key)
         }
-        let state = target._s
         let data = { target }
-        if (state) {
-          let resolved = state.applyContext(target._sc)
-          if (resolved) {
-            target._s = state = resolved
-            target._sc = state.storeContext()
-          } else if (resolved === null) {
-            target._s = null
-            delete target._sc
-            state = null
-          }
-          data.state = state
-        }
+        restore(data)
         listener.emit(elem, attach(e, data), stamp)
         if (data.prevent) {
           break
